requirements.txt
boto3
matplotlib
pyyaml
hjson
